+++ internal/ui/views/connection.go
@@ -4,41 +4,27 @@ import (
 	"github.com/thomas-marquis/s3-box/internal/domain/connection_deck"
 	appcontext "github.com/thomas-marquis/s3-box/internal/ui/app/context"
 	"github.com/thomas-marquis/s3-box/internal/ui/app/navigation"
-	"github.com/thomas-marquis/s3-box/internal/ui/views/components"
+	"github.com/thomas-marquis/s3-box/internal/ui/views/widget"
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/container"
-	"fyne.io/fyne/v2/data/binding"
 	"fyne.io/fyne/v2/theme"
-	"fyne.io/fyne/v2/widget"
+	fyne_widget "fyne.io/fyne/v2/widget"
 )
 
 func GetConnectionView(appCtx appcontext.AppContext) (*fyne.Container, error) {
-	deck := appCtx.ConnectionViewModel().Deck()
-	connLine := components.NewConnectionLine(deck)
-	connectionsList := widget.NewListWithData(
-		appCtx.ConnectionViewModel().Connections(),
-		func() fyne.CanvasObject {
-			return connLine.Raw()
-		},
-		func(di binding.DataItem, obj fyne.CanvasObject) {
-			i, _ := di.(binding.Untyped).Get()
-			conn, _ := i.(*connection_deck.Connection)
-			connLine.Update(appCtx, obj, conn)
-		},
-	)
-	createBtn := widget.NewButtonWithIcon(
+	connectionsList := widget.NewConnectionList(appCtx)
+
+	createBtn := fyne_widget.NewButtonWithIcon(
 		"New connection",
 		theme.ContentAddIcon(),
-		func() {
-			components.NewConnectionDialog(appCtx,
-				"New connection",
-				connection_deck.Connection{},
-				false,
-				appCtx.ConnectionViewModel().Create).Show()
-		})
+		widget.NewConnectionForm(appCtx,
+			&connection_deck.Connection{},
+			false,
+			appCtx.ConnectionViewModel().Create,
+		).AsDialog("New connection").Show)
 
-	goToExplorerBtn := widget.NewButtonWithIcon(
+	goToExplorerBtn := fyne_widget.NewButtonWithIcon(
 		"View files",
 		theme.NavigateBackIcon(),
 		func() {
@@ -46,12 +32,11 @@ func GetConnectionView(appCtx appcontext.AppContext) (*fyne.Container, error) {
 		},
 	)
 
-	mainContainer := container.NewBorder(
+	return container.NewBorder(
 		container.NewHBox(goToExplorerBtn),
 		container.NewCenter(createBtn),
 		nil,
 		nil,
 		connectionsList,
-	)
-	return mainContainer, nil
+	), nil
 }
